requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-community==0.0.24
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf2==3.0.1
python-docx==1.1.0
ollama==0.1.6
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
plotly==5.18.0
